requirements.txt
bleach==2.1.2
Cerberus==0.9.2
certifi==2017.11.5
chardet==3.0.4
click==6.7
decorator==4.2.1
entrypoints==0.2.3
Eve==0.7.6
Events==0.2.2
Flask==0.12
Flask-PyMongo==0.5.1
html5lib==1.0.1
idna==2.6
ipykernel==4.7.0
ipython==6.2.1
ipython-genutils==0.2.0
ipywidgets==7.1.0
itsdangerous==0.24
jedi==0.11.1
Jinja2==2.10
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==5.2.1
jupyter-console==5.2.0
jupyter-core==4.4.0
MarkupSafe==0.23
mistune==0.8.3
nbconvert==5.3.1
nbformat==4.4.0
notebook==5.2.2
numpy==1.14.0
pandas==0.22.0
pandocfilters==1.4.2
parso==0.1.1
pexpect==4.3.1
pickleshare==0.7.4
pkg-resources==0.0.0
prompt-toolkit==1.0.15
ptyprocess==0.5.2
Pygments==2.2.0
pymongo==3.6.0
python-dateutil==2.6.1
python-dotenv==0.7.1
pytz==2017.3
pyzmq==16.0.3
qtconsole==4.3.1
requests==2.18.4
scipy==1.0.0
simplegeneric==0.8.1
simplejson==3.13.2
six==1.11.0
terminado==0.8.1
testpath==0.3.1
tornado==4.5.3
traitlets==4.3.2
urllib3==1.22
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.11.15
widgetsnbextension==3.1.0
